[EXPERIMENTAL] Fix contexts when interrupts occur in a callgate
THIS CODE IS UNTESTED (because I don't have any simple trick to
deterministically run the corresponding code branch)

This commit tries to fix the unfortunate race condition on a callgate
`cli`. The *untested* code jumps to the stack base and retrieves missing
infos pushed by the callgate in order to fix the malformed context.
The assembly code moves CS, EIP and fetches SS and ESP from the stack
bottom. It also fixes EFLAGS by setting back the interrupt flag, and
fixes EIP so that the far call would execute again when the context is
resumed.