Fix expected stack layout when trying to detect a race condition
Previous iteration of assembly code meant to detect an interrupt during a
farcall used to use the `call` instruction to run the detection code,
pushing a return address on the stack. This mechanism has been superseded
by a macro placed in a specific execution path to prevent the other
(unaffected) execution paths to run the check.
But I forgot to remove the return address from the stack layout in the
process, my bad.