New configuration script version
This new configuration script has three mandatory arguments which are "--architecture",
"--partition-name" and "--libpip" as well as optional arguments. It will try to detect
if the required tools are installed and if they have the right version. If all the tools
are valid, it generates the "/path/to/pipcore/toolchain.mk" and
"/path/to/pipcore/src/arch/$architecture/partitions/toolchain.mk" configuration files.
For more information, run "./configure --help".